requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
DateTime==5.5
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.0.2
oauth==1.0.1
pandas==2.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.21.0
zope.interface==7.2
